requirements.txt
streamlit>=1.35.0
google-generativeai>=0.5.0
langchain>=0.2.0
langchain-google-genai>=1.0.0
faiss-cpu>=1.8.0
pdfplumber>=0.11.0
PyPDF2>=3.0.0
python-docx>=1.1.0
python-pptx>=0.6.23
numpy>=1.26.0
python-dotenv>=1.0.0
